[chore] Move back processor definitions, make profile embed processor
