--- processor/internal/metrics.go
@@ -1,15 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package internal // import "go.opentelemetry.io/collector/processor/internal"
-
-import (
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/consumer"
-)
-
-// Metrics is a processor that can consume metrics.
-type Metrics interface {
-	component.Component
-	consumer.Metrics
-}
