+++ processor/processorprofiles/processor.go
@@ -4,18 +4,124 @@
 package processorprofiles // import "go.opentelemetry.io/collector/processor/processorprofiles"
 
 import (
+	"context"
+
 	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/consumer/consumerprofiles"
+	"go.opentelemetry.io/collector/pipeline"
 	"go.opentelemetry.io/collector/processor"
-	"go.opentelemetry.io/collector/processor/internal"
 )
 
+// Factory is a component.Factory interface for processors.
+//
+// This interface cannot be directly implemented. Implementations must
+// use the NewFactory to implement it.
+type Factory interface {
+	processor.Factory
+
+	// CreateProfilesProcessor creates a ProfilesProcessor based on this config.
+	// If the processor type does not support tracing or if the config is not valid,
+	// an error will be returned instead.
+	CreateProfilesProcessor(ctx context.Context, set processor.Settings, cfg component.Config, nextConsumer consumerprofiles.Profiles) (Profiles, error)
+
+	// ProfilesProcessorStability gets the stability level of the ProfilesProcessor.
+	ProfilesProcessorStability() component.StabilityLevel
+}
+
 // Profiles is a processor that can consume profiles.
-type Profiles = internal.Profiles
+type Profiles interface {
+	component.Component
+	consumerprofiles.Profiles
+}
 
 // CreateProfilesFunc is the equivalent of Factory.CreateProfiles().
-type CreateProfilesFunc = internal.CreateProfilesFunc
+// CreateProfilesFunc is the equivalent of Factory.CreateProfiles().
+type CreateProfilesFunc func(context.Context, processor.Settings, component.Config, consumerprofiles.Profiles) (Profiles, error)
+
+// CreateProfilesProcessor implements Factory.CreateProfilesProcessor().
+func (f CreateProfilesFunc) CreateProfilesProcessor(
+	ctx context.Context,
+	set processor.Settings,
+	cfg component.Config,
+	nextConsumer consumerprofiles.Profiles) (Profiles, error) {
+	if f == nil {
+		return nil, pipeline.ErrSignalNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
+
+// FactoryOption apply changes to ReceiverOptions.
+type FactoryOption interface {
+	// applyOption applies the option.
+	applyOption(o *factoryOpts)
+}
+
+// factoryOptionFunc is an ReceiverFactoryOption created through a function.
+type factoryOptionFunc func(*factoryOpts)
+
+func (f factoryOptionFunc) applyOption(o *factoryOpts) {
+	f(o)
+}
+
+type factory struct {
+	processor.Factory
+	CreateProfilesFunc
+	profilesStabilityLevel component.StabilityLevel
+}
+
+func (f factory) ProfilesProcessorStability() component.StabilityLevel {
+	return f.profilesStabilityLevel
+}
+
+type factoryOpts struct {
+	cfgType component.Type
+	component.CreateDefaultConfigFunc
+	opts []processor.FactoryOption
+	CreateProfilesFunc
+	profilesStabilityLevel component.StabilityLevel
+}
+
+// WithTraces overrides the default "error not supported" implementation for CreateTraces and the default "undefined" stability level.
+func WithTraces(createTraces processor.CreateTracesFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factoryOpts) {
+		o.opts = append(o.opts, processor.WithTraces(createTraces, sl))
+	})
+}
+
+// WithMetrics overrides the default "error not supported" implementation for CreateMetrics and the default "undefined" stability level.
+func WithMetrics(createMetrics processor.CreateMetricsFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factoryOpts) {
+		o.opts = append(o.opts, processor.WithMetrics(createMetrics, sl))
+	})
+}
+
+// WithLogs overrides the default "error not supported" implementation for CreateLogs and the default "undefined" stability level.
+func WithLogs(createLogs processor.CreateLogsFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factoryOpts) {
+		o.opts = append(o.opts, processor.WithLogs(createLogs, sl))
+	})
+}
 
 // WithProfiles overrides the default "error not supported" implementation for CreateProfiles and the default "undefined" stability level.
-func WithProfiles(createProfiles CreateProfilesFunc, sl component.StabilityLevel) processor.FactoryOption {
-	return internal.WithProfiles(createProfiles, sl)
+func WithProfiles(createProfiles CreateProfilesFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factoryOpts) {
+		o.profilesStabilityLevel = sl
+		o.CreateProfilesFunc = createProfiles
+	})
+}
+
+// NewFactory returns a Factory.
+func NewFactory(cfgType component.Type, createDefaultConfig component.CreateDefaultConfigFunc, options ...FactoryOption) Factory {
+	opts := factoryOpts{
+		cfgType:                 cfgType,
+		CreateDefaultConfigFunc: createDefaultConfig,
+	}
+	for _, opt := range options {
+		opt.applyOption(&opts)
+	}
+	return &factory{
+		Factory:                processor.NewFactory(opts.cfgType, opts.CreateDefaultConfig, opts.opts...),
+		CreateProfilesFunc:     opts.CreateProfilesFunc,
+		profilesStabilityLevel: opts.profilesStabilityLevel,
+	}
 }
