+++ service/internal/builders/processor_test.go
@@ -24,12 +24,12 @@ func TestProcessorBuilder(t *testing.T) {
 	defaultCfg := struct{}{}
 	factories, err := processor.MakeFactoryMap([]processor.Factory{
 		processor.NewFactory(component.MustNewType("err"), nil),
-		processor.NewFactory(
+		processorprofiles.NewFactory(
 			component.MustNewType("all"),
 			func() component.Config { return &defaultCfg },
-			processor.WithTraces(createProcessorTraces, component.StabilityLevelDevelopment),
-			processor.WithMetrics(createProcessorMetrics, component.StabilityLevelAlpha),
-			processor.WithLogs(createProcessorLogs, component.StabilityLevelDeprecated),
+			processorprofiles.WithTraces(createProcessorTraces, component.StabilityLevelDevelopment),
+			processorprofiles.WithMetrics(createProcessorMetrics, component.StabilityLevelAlpha),
+			processorprofiles.WithLogs(createProcessorLogs, component.StabilityLevelDeprecated),
 			processorprofiles.WithProfiles(createProcessorProfiles, component.StabilityLevelDevelopment),
 		),
 	}...)
@@ -136,12 +136,12 @@ func TestProcessorBuilder(t *testing.T) {
 func TestProcessorBuilderMissingConfig(t *testing.T) {
 	defaultCfg := struct{}{}
 	factories, err := processor.MakeFactoryMap([]processor.Factory{
-		processor.NewFactory(
+		processorprofiles.NewFactory(
 			component.MustNewType("all"),
 			func() component.Config { return &defaultCfg },
-			processor.WithTraces(createProcessorTraces, component.StabilityLevelDevelopment),
-			processor.WithMetrics(createProcessorMetrics, component.StabilityLevelAlpha),
-			processor.WithLogs(createProcessorLogs, component.StabilityLevelDeprecated),
+			processorprofiles.WithTraces(createProcessorTraces, component.StabilityLevelDevelopment),
+			processorprofiles.WithMetrics(createProcessorMetrics, component.StabilityLevelAlpha),
+			processorprofiles.WithLogs(createProcessorLogs, component.StabilityLevelDeprecated),
 			processorprofiles.WithProfiles(createProcessorProfiles, component.StabilityLevelDevelopment),
 		),
 	}...)
@@ -207,7 +207,7 @@ func TestNewNopProcessorBuilder(t *testing.T) {
 	require.NoError(t, err)
 	assert.IsType(t, logs, bLogs)
 
-	profiles, err := factory.CreateProfilesProcessor(context.Background(), set, cfg, consumertest.NewNop())
+	profiles, err := factory.(processorprofiles.Factory).CreateProfilesProcessor(context.Background(), set, cfg, consumertest.NewNop())
 	require.NoError(t, err)
 	bProfiles, err := builder.CreateProfiles(context.Background(), set, consumertest.NewNop())
 	require.NoError(t, err)
